move chain metrics back to smartengine to make it reusable
