+++ crates/fluvio-smartengine/Cargo.toml
@@ -13,6 +13,7 @@ description = "The official Fluvio SmartEngine"
 
 [features]
 wasi = ["wasmtime-wasi"]
+metrics = ["serde"]
 
 
 [dependencies]
@@ -21,6 +22,7 @@ thiserror = "1"
 anyhow = { version = "1.0.38" }
 wasmtime = { version = "2.0.0"}
 wasmtime-wasi = { version = "2.0.0", optional = true}
+serde = { version = "1.0.103", features = ['derive'], optional = true }
 
 cfg-if = "1.0.0"
 derive_builder = "0.11.0"
