+++ crates/fluvio-smartengine/src/lib.rs
@@ -7,6 +7,9 @@ mod engine;
 pub use engine::*;
 pub mod instance;
 
+#[cfg(feature = "metrics")]
+pub mod metrics;
+
 pub type WasmSlice = (i32, i32, u32);
 pub type Version = i16;
