+++ crates/fluvio-spu/Cargo.toml
@@ -67,7 +67,7 @@ fluvio-future = { version = "0.4.0", features = [
     "openssl_tls",
     "zero_copy",
 ] }
-fluvio-smartengine = { path = "../fluvio-smartengine" }
+fluvio-smartengine = { path = "../fluvio-smartengine", features = ["metrics"]}
 fluvio-smartmodule = { path = "../fluvio-smartmodule",  default-features = false }
 
 [dev-dependencies]
