+++ crates/fluvio-spu/src/core/metrics.rs
@@ -2,6 +2,7 @@ use std::sync::{
     atomic::{AtomicU64, Ordering},
 };
 
+use fluvio_smartengine::metrics::SmartModuleChainMetrics;
 use serde::Serialize;
 
 #[derive(Default, Debug, Serialize)]
@@ -60,23 +61,5 @@ impl<'a> SpuMetricsTopicPartition<'a> {
         self.metrics
             .bytes_written
             .fetch_add(value, Ordering::SeqCst);
-    }
-}
-
-#[derive(Serialize, Default, Debug)]
-pub struct SmartModuleChainMetrics {
-    bytes_in: AtomicU64,
-    records_out: AtomicU64,
-    invocation_count: AtomicU64,
-}
-
-impl SmartModuleChainMetrics {
-    pub(crate) fn add_bytes_in(&self, value: u64) {
-        self.bytes_in.fetch_add(value, Ordering::SeqCst);
-        self.invocation_count.fetch_add(1, Ordering::SeqCst);
-    }
-
-    pub(crate) fn add_records_out(&self, value: u64) {
-        self.records_out.fetch_add(value, Ordering::SeqCst);
     }
 }
